#18 fix([CLI]): loading indicator glitches removed
